refactor: Car객체를 Cars 클래스 내부에서 생성하도록 수정
