+++ src/main/java/racingcar/controller/RacingCarGame.java
@@ -1,45 +1,24 @@
 package racingcar.controller;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import racingcar.model.Car;
 import racingcar.model.Cars;
 import racingcar.model.GameTurn;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
 public class RacingCarGame {
-	private static final String COMMA_DELIMITER = ",";
-
 	private final InputView inputView = new InputView();
 	private final OutputView outputView = new OutputView();
 	private Cars cars;
 	private GameTurn gameTurn;
 
 	public void makeCars() {
-		cars = new Cars(toCar(getCarNames()));
-	}
-
-	private List<Car> toCar(List<String> carNames) {
+		cars = new Cars(null);
 		try {
-			return carNames.stream()
-				.map(Car::new)
-				.collect(Collectors.toList());
+			cars.makeCars(inputView.getCarNames());
 		} catch (IllegalArgumentException exception) {
 			System.out.println(exception.getMessage());
-			return toCar(getCarNames());
+			makeCars();
 		}
-	}
-
-	private List<String> getCarNames() {
-		String inputCarNames = inputView.getCarNames();
-		return splitCarNames(inputCarNames);
-	}
-
-	private List<String> splitCarNames(String carNames) {
-		return Arrays.asList(carNames.split(COMMA_DELIMITER));
 	}
 
 	public void saveGameTurn() {
