+++ src/main/java/racingcar/model/Cars.java
@@ -1,16 +1,33 @@
 package racingcar.model;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import racingcar.service.RandomNumberService;
 
 public class Cars {
-	private final List<Car> cars;
+	private static final String COMMA_DELIMITER = ",";
+
+	private List<Car> cars;
 
 	public Cars(List<Car> cars) {
 		this.cars = cars;
+	}
+
+	public void makeCars(String carNames) {
+		this.cars = toCar(splitCarNames(carNames));
+	}
+
+	private List<Car> toCar(List<String> carNames) {
+		return carNames.stream()
+			.map(Car::new)
+			.collect(Collectors.toList());
+	}
+
+	private List<String> splitCarNames(String carNames) {
+		return Arrays.asList(carNames.split(COMMA_DELIMITER));
 	}
 
 	public void moveCars() {
