+++ src/test/java/racingcar/RacingCarTest.java
@@ -42,13 +42,7 @@ public class RacingCarTest {
 		assertThatThrownBy(() -> new GameTurn("-1"))
 			.isInstanceOf(IllegalArgumentException.class);
 	}
-
-	// @Test
-	// public void 자동차_객체_생성() throws Exception {
-	// 	List<Car> result = racingCarGame.toCar(Arrays.asList("배카라", "아스피"));
-	// 	assertThat(result).isEqualTo(Arrays.asList(new Car("배카라"), new Car("아스피")));
-	// }
-
+	
 	@Test
 	public void 전진_여부() {
 		Car testCar = new Car("아스피", 2);
